[api] Add native TLS support for redis
Add use_ssl, ca options to app JSON config for redis. Default to system
certs if CA not specified.

* https://docs.redislabs.com/latest/rs/references/client_references/client_ioredis/
* Move LINUX_CERTFILES const to config/conf.js
* export configureCA() from config/conf.js
