+++ opencti-platform/opencti-graphql/src/database/redis.js
@@ -1,8 +1,9 @@
+import { readFileSync } from 'fs';
 import Redis from 'ioredis';
 import Redlock from 'redlock';
 import { RedisPubSub } from 'graphql-redis-subscriptions';
 import * as R from 'ramda';
-import conf, { logApp } from '../config/conf';
+import conf, { configureCA, logApp } from '../config/conf';
 import {
   generateLogMessage,
   isEmptyField,
@@ -24,16 +25,21 @@ import { now } from '../utils/format';
 import RedisStore from './sessionStore-redis';
 import SessionStoreMemory from './sessionStore-memory';
 
+const USE_SSL = conf.get('redis:use_ssl');
+const REDIS_CA = conf.get('redis:ca').map((path) => readFileSync(path));
+
 const BASE_DATABASE = 0; // works key for tracking / stream
 export const CONTEXT_DATABASE = 1; // locks / user context
 export const SESSION_DATABASE = 2; // locks / user context
 const OPENCTI_STREAM = 'stream.opencti';
 const REDIS_EXPIRE_TIME = 90;
+
 const redisOptions = (database) => ({
   lazyConnect: true,
   db: database,
   port: conf.get('redis:port'),
   host: conf.get('redis:hostname'),
+  tls: USE_SSL ? configureCA(REDIS_CA) : null,
   username: conf.get('redis:username'),
   password: conf.get('redis:password'),
   retryStrategy: /* istanbul ignore next */ (times) => Math.min(times * 50, 2000),
